Refactoring and clean up
